requirements.txt
Django>=4.2,<5.1
djangorestframework>=3.14
django-cors-headers>=4.3
channels>=4.0
daphne>=4.0
whitenoise>=6.6
